fix survey information
The index was modified in https://github.com/ScienceParkStudyGroup/2022-04-21-Python-workshop/commit/5bec0748c9f4ec0a0b33fec93efa693ae59fe128, which brought in the structure from an older version of the workshop template that contained outdated URLs for the surveys (which were updated in October 2019 https://github.com/carpentries/workshop-template/pull/628). 

This PR fixes the survey links so that the should point to the correct places.